Move the metadata cops to ChefModernize
These cops really aren't deprecations. They're just not necessary
anymore. We shouldn't be alerting in chef-analyze that every customer
needs to drop everything and remove long_description from their
metadata. It doesn't warrant that, but if they're already in a cookbook
they should def clean it up.
