+++ lib/rubocop/cop/chef/modernize/provides_metadata.rb
@@ -18,20 +18,20 @@
 module RuboCop
   module Cop
     module Chef
-      module ChefDeprecations
-        # Don't use the deprecated 'suggests' metadata value
+      module ChefModernize
+        # The provides metadata.rb method is not used and is unnecessary in cookbooks.
         #
         # @example
         #
         #   # bad in metadata.rb:
         #
-        #   suggests "another_cookbook"
+        #   provides "some_thing"
         #
-        class SuggestsMetadata < Cop
-          MSG = "Don't use the deprecated 'suggests' metadata value".freeze
+        class ProvidesMetadata < Cop
+          MSG = "The provides metadata.rb method is not used and is unnecessary in cookbooks.".freeze
 
           def on_send(node)
-            add_offense(node, location: :expression, message: MSG, severity: :refactor) if node.method_name == :suggests
+            add_offense(node, location: :expression, message: MSG, severity: :refactor) if node.method_name == :provides
           end
 
           def autocorrect(node)
