+++ lib/rubocop/cop/chef/modernize/respond_to_resource_name.rb
@@ -18,9 +18,7 @@ module RuboCop
   module Cop
     module Chef
       module ChefModernize
-        # Chef 12.5 introduced the resource_name method for resources. Many cookbooks used
-        # respond_to?(:resource_name) to provide backwards compatibility with older chef-client
-        # releases. This backwards compatibility is no longer necessary.
+        # Chef Infra Client 12.5 introduced the resource_name method for resources. Many cookbooks used respond_to?(:resource_name) to provide backwards compatibility with older chef-client releases. This backwards compatibility is no longer necessary.
         #
         # @example
         #
